requirements.txt
-e .
aws-cdk.aws-ec2
aws-cdk.aws-s3
aws-cdk.aws-ecr
aws-cdk.aws-ecs
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-cloudfront
aws-cdk.aws-ecr-assets
aws-cdk.aws-rds
aws-cdk.aws-elasticache